some nits; made `GetBaseline` a util function
