+++ protocol/x/ratelimit/module.go
@@ -153,7 +153,7 @@ func (AppModule) ConsensusVersion() uint64 { return 1 }
 // EndBlock contains the logic that is automatically triggered at the end of each block
 func (am AppModule) EndBlock(ctx context.Context) error {
 	sdkCtx := sdk.UnwrapSDKContext(ctx)
-	am.keeper.UpdateCapacityEndBlocker(sdkCtx)
+	am.keeper.UpdateAllCapacitiesEndBlocker(sdkCtx)
 
 	return nil
 }
